pyproject.toml
[project]
name = "osstriage"
version = "0.1.0"
description = "AI-driven CLI & GitHub Action for automated Pull Request review and Issue triaging."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
authors = [
    { name = "OSSTriage Contributors" },
]
keywords = ["github", "pull-request", "code-review", "ai", "dspy", "issue-triage", "oss"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "dspy>=2.6",
    "PyGithub>=2.5",
    "httpx>=0.28",
    "typer>=0.15",
    "rich>=13.9",
    "python-dotenv>=1.1",
]

[project.scripts]
osstriage = "osstriage.cli:app"

[project.urls]
Homepage = "https://github.com/your-org/osstriage"
Repository = "https://github.com/your-org/osstriage"
Issues = "https://github.com/your-org/osstriage/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/osstriage"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
